fix: Overwriting clock in
change: I used the enumerate method to get the idex of the current date's row.
issue: The first row of the id's clock in time was overwritten instead of today's clock in time.
reason: I used the find method of the gspread package and it returns the first matching cell.